lz4 decompress
	geändert:       Makefile
	geändert:       lz4mt.c
	geändert:       lz4mt.h
	geändert:       lz4mt_compress.c
	geändert:       lz4mt_decompress.c
	geändert:       lz5mt.c
	geändert:       lz5mt.h
	geändert:       lz5mt_compress.c
	geändert:       zstdmt.h
	geändert:       zstdmt_compress.c
	geändert:       zstdmt_decompress.c

+++ zstdmt.h
@@ -30,7 +30,7 @@ extern "C" {
 
 typedef struct {
 	void *buf;		/* ptr to data */
-	int size;		/* length of buf */
+	size_t size;		/* length of buf */
 } ZSTDMT_Buffer;
 
 /**
